[feature] Implement inventory management modules
This commit implements the inventory management modules for controller and peer nodes. Peers are kept stupid and only report the UID and item name of scanned products. Central node handles transaction direction logic and logs transactions, updates inventory records, and sends alerts.

+++ Core/Inc/drivers/mfrc522.h
@@ -225,7 +225,7 @@ uint8_t mfrc_picc_write(uint8_t picc_block_addr, uint8_t *send_data);
  * @param picc_addr_block addres of the block to read from
  * @param rcv_data buffer into which data will be read
  */
-uint8_t mfrc_picc_read(uint8_t picc_block_addr, uint8_t *rcv_data);
+uint16_t mfrc_picc_read(uint8_t picc_block_addr, uint8_t *rcv_data);
 
 #endif
